refactor: move server/backends tests into test/server/backends
