+++ test/server/backends/worklog.spec.ts
@@ -1,6 +1,6 @@
 import { describe, it, expect, vi } from "vitest";
 import { SCHEDULE_TYPES } from "@receptron/task-scheduler";
-import { worklogSystemTask, WORKLOG_PROMPT } from "./worklog.js";
+import { worklogSystemTask, WORKLOG_PROMPT } from "../../../../server/../../server/backends/worklog.js";
 
 const HOUR_MS = 3_600_000;
