+++ test/server/backends/workspaceSetup.spec.ts
@@ -3,7 +3,7 @@ import { describe, it, expect, beforeEach, afterEach } from "vitest";
 import { mkdtempSync, rmSync, existsSync, readdirSync } from "node:fs";
 import { tmpdir, homedir } from "node:os";
 import path from "node:path";
-import { isManagedWorkspace, initWorkspaceSetup } from "./workspaceSetup.js";
+import { isManagedWorkspace, initWorkspaceSetup } from "../../../../server/../../server/backends/workspaceSetup.js";
 
 // Save/restore MULMOCLAUDE_WORKSPACE_PATH so a test can mark a temp dir as the
 // managed workspace without leaking into sibling tests.
